refactor: profile사용을 위한 분리
